Add files via upload
This is a project where I have tried to code the American_sign_language_detection project. Here firstly I have tried to train my model with the help of the free GPU of google colab. My model has reached a 0.86 test accuracy which is a reasonable good to work with, then I have imported the JSON file and the h5 file where my model consists of all the weights and biases of the model. Then I implement the model with my local Jupiter notebook and run this awesome project.